✅ Securing the App

+++ app/api/issues/[id]/route.ts
@@ -1,12 +1,14 @@
 import { issueSchema } from "@/app/validationSchema";
 import prisma from "@/prisma/db";
-import delay from "delay";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
 interface IParams {
   params: { id: string };
 }
 export async function PATCH(request: Request, { params: { id } }: IParams) {
+  const session = await getServerSession();
+  if (!session) return NextResponse.json({}, { status: 401 });
   try {
     const body = await request.json();
     const validation = issueSchema.safeParse(body);
@@ -33,16 +35,16 @@ export async function PATCH(request: Request, { params: { id } }: IParams) {
   }
 }
 export async function DELETE(request: Request, { params: { id } }: IParams) {
+  const session = await getServerSession();
+  if (!session) return NextResponse.json({}, { status: 401 });
   try {
     const issue = await prisma.issue.findUnique({
       where: { id: parseInt(id) },
     });
     if (!issue) return new NextResponse("Invalid ID", { status: 404 });
 
-
     await prisma.issue.delete({ where: { id: parseInt(id) } });
     return NextResponse.json({});
-    
   } catch (error: any) {
     console.log(error, "Error Message");
     return new NextResponse("Internal Error", { status: 500 });
